requirements_macos.txt
PyYAML==6.0
rumps==0.4.0
pyinstaller
requests
# Pillow is needed for pyinstaller to be able to convert .png image to .icns to allow to use as macOS file icon
Pillow
mac-login-items
